Refactor HTTP-01 listener.
