+++ certcache.go
@@ -3,8 +3,11 @@ package nitriding
 import (
 	"context"
 	"errors"
+	"net"
+	"net/http"
 	"sync"
 
+	"github.com/mdlayher/vsock"
 	"golang.org/x/crypto/acme/autocert"
 )
 
@@ -49,4 +52,32 @@ func (c *certCache) Delete(ctx context.Context, key string) error {
 
 	delete(c.cache, key)
 	return nil
+}
+
+func listenHTTP01(errChan chan error, mgr *autocert.Manager) {
+	// Let's Encrypt's HTTP-01 challenge requires a listener on port 80:
+	// https://letsencrypt.org/docs/challenge-types/#http-01-challenge
+	var l net.Listener
+	var err error
+
+	if inEnclave {
+		l, err = vsock.Listen(uint32(80), nil)
+		if err != nil {
+			errChan <- errHTTP01Failed
+			return
+		}
+		defer func() {
+			_ = l.Close()
+		}()
+	} else {
+		l, err = net.Listen("tcp", ":80")
+		if err != nil {
+			errChan <- errHTTP01Failed
+			return
+		}
+	}
+
+	elog.Print("Starting autocert listener.")
+	errChan <- nil
+	_ = http.Serve(l, mgr.HTTPHandler(nil))
 }
